+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Methods to generate a new token from a member object or his credentials (unique_identifier and password)
+- Made the included member fields maintainable through `included_member_fields` config
 
 ## [0.1.0]
 ### Added